Issue 154: Move GiraffaConfiguration to hbase package. (Milan)
