+++ giraffa-core/src/main/java/org/apache/giraffa/GiraffaFileSystem.java
@@ -23,6 +23,7 @@ import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.EnumSet;
 
+import org.apache.giraffa.hbase.GiraffaConfiguration;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.ContentSummary;
 import org.apache.hadoop.fs.CreateFlag;
