+++ giraffa-core/src/main/java/org/apache/giraffa/GiraffaFs.java
@@ -17,6 +17,7 @@
  */
 package org.apache.giraffa;
 
+import org.apache.giraffa.hbase.GiraffaConfiguration;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.DelegateToFileSystem;
