+++ giraffa-core/src/main/java/org/apache/giraffa/hbase/INodeManager.java
@@ -25,7 +25,6 @@ import java.util.List;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.giraffa.FileField;
-import org.apache.giraffa.GiraffaConfiguration;
 import org.apache.giraffa.INode;
 import org.apache.giraffa.RowKey;
 import org.apache.giraffa.RowKeyBytes;
