+++ giraffa-core/src/main/java/org/apache/giraffa/hbase/NamespaceAgent.java
@@ -47,7 +47,6 @@ import java.util.List;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.giraffa.FileField;
-import org.apache.giraffa.GiraffaConfiguration;
 import org.apache.giraffa.NamespaceService;
 import org.apache.giraffa.RowKey;
 import org.apache.giraffa.RowKeyFactory;
