+++ giraffa-core/src/main/resources/hbase-webapps/giraffa/htable.jsp
@@ -4,7 +4,7 @@
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 <%@ page import="org.apache.hadoop.hbase.client.HBaseAdmin" %>
 <%@ page import="org.apache.hadoop.conf.Configuration" %>
-<%@ page import="org.apache.giraffa.GiraffaConfiguration" %>
+<%@ page import="org.apache.giraffa.hbase.GiraffaConfiguration" %>
 
 
 <%
