+++ giraffa-core/src/main/resources/hbase-webapps/giraffa/index.jsp
@@ -3,7 +3,7 @@
 <%@taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 <%@ page import="org.apache.hadoop.hbase.client.HBaseAdmin" %>
 <%@ page import="org.apache.hadoop.conf.Configuration" %>
-<%@ page import="org.apache.giraffa.GiraffaConfiguration" %>
+<%@ page import="org.apache.giraffa.hbase.GiraffaConfiguration" %>
 <%@ page import="org.apache.hadoop.hbase.io.ImmutableBytesWritable" %>
 <%@ page import="org.apache.giraffa.RowKeyBytes" %>
 <%@ page import="org.apache.hadoop.hbase.HTableDescriptor" %>
