+++ giraffa-core/src/test/java/org/apache/giraffa/TestGiraffaFS.java
@@ -18,6 +18,8 @@
 package org.apache.giraffa;
 
 import java.io.IOException;
+
+import org.apache.giraffa.hbase.GiraffaConfiguration;
 import org.apache.hadoop.fs.FileAlreadyExistsException;
 import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.FileSystem;
