+++ giraffa-core/src/test/java/org/apache/giraffa/TestGiraffaFSContract.java
@@ -23,6 +23,7 @@ import junit.extensions.TestSetup;
 import junit.framework.Test;
 import junit.framework.TestSuite;
 
+import org.apache.giraffa.hbase.GiraffaConfiguration;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.FileSystemContractBaseTest;
 import org.apache.hadoop.hbase.HBaseTestingUtility;
