+++ giraffa-core/src/test/java/org/apache/giraffa/TestGiraffaFileStatus.java
@@ -18,6 +18,8 @@
 package org.apache.giraffa;
 
 import java.io.IOException;
+
+import org.apache.giraffa.hbase.GiraffaConfiguration;
 import org.apache.hadoop.fs.FSDataOutputStream;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
