+++ giraffa-core/src/test/java/org/apache/giraffa/TestGiraffaUpgrade.java
@@ -34,6 +34,7 @@ import java.util.Date;
 import java.util.List;
 
 import org.apache.giraffa.GiraffaConstants.FileState;
+import org.apache.giraffa.hbase.GiraffaConfiguration;
 import org.apache.giraffa.hbase.INodeManager;
 import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.FileSystem;
