+++ giraffa-core/src/test/java/org/apache/giraffa/TestLeaseManagement.java
@@ -29,6 +29,7 @@ import static org.junit.Assert.fail;
 import java.io.IOException;
 import java.util.Collection;
 
+import org.apache.giraffa.hbase.GiraffaConfiguration;
 import org.apache.giraffa.hbase.INodeManager;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FSDataOutputStream;
